Change the generation of ID for games to have it set by the persistance layout

+++ src/main/java/ca/ulaval/glo4003/domain/services/SportService.java
@@ -64,7 +64,7 @@ public class SportService {
 	private void countNumberOfTickets(List<GameDto> games) throws GameDoesntExistException {
 		for (GameDto game : games) {
 			try {
-				game.setNumberOfTickets(ticketDao.getTicketsForGame((int) game.getId()).size());
+				game.setNumberOfTickets(ticketDao.getTicketsForGame(game.getId()).size());
 			} catch (Exception e) {
 				game.setNumberOfTickets(0);
 			}
